t5310: stop expecting the default branch name `master`
To open the door for changing the fall-back value for
`init.defaultBranch` when it is not set by the user, we started to
modify the test suite to work with any such fall-back value (instead of
expecting the default branch name to be `master`).

To that end, we specifically rename the initial branch from whatever it
was called to `second` (forcing it, just in case that the hard-coded
fall-back is `second`, in which case the rename is a no-op that would
fail unless forced).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>